need to implement Conversation History objects
