--- captions.json
@@ -1,18 +1,0 @@
-[
-  {
-    "path": "/home/darko/workspace/assets/cap_img/image4.png",
-    "caption": "Close-up view of a Toyota Tundra's front grille covered in ice. The large \"TOYOTA\" lettering is prominently displayed on the black honeycomb-patterned grille. Icicles have formed on the grille's edges and between the hexagonal openings, creating a frosty, textured appearance against the dark background."
-  },
-  {
-    "path": "/home/darko/workspace/assets/cap_img/image1.png",
-    "caption": "A serene lake reflects a forest of golden-lit evergreen trees under a dramatic cloudy sky. The shoreline is fringed with lush green grass, creating a tranquil scene where water meets land. Sunlight breaking through the clouds illuminates the trees, contrasting with the moody atmosphere above."
-  },
-  {
-    "path": "/home/darko/workspace/assets/cap_img/image2.png",
-    "caption": "A stainless steel Breville espresso machine sits on a marble countertop. The machine features multiple blue-lit buttons, a pressure gauge, and a built-in grinder on top filled with coffee beans. A white coffee mug rests on the machine's cup warmer. The sleek design and professional appearance suggest this is a high-end home espresso maker."
-  },
-  {
-    "path": "/home/darko/workspace/assets/cap_img/image3.png",
-    "caption": "A gray tabby cat sleeps peacefully on a plush white blanket or pet bed. The cat's eyes are closed, and its pink tongue is slightly visible, giving it a contented expression. The cozy scene is set against a dark background, emphasizing the comfort of the cat's resting spot."
-  }
-]
